requirements.txt
discord.py
flask
flask-cors
python-dotenv
sqlalchemy
psutil
requests
